implement pendulum state machine. some features are not used anymore
pendulum position is just a linear function of time. it used to be a sine
function...
time precision is worse, because we use milliseconds now.
error level and watchdog are not used anymore, because we have states now.

P_RESET does not work too well to get the pendulum going